Fixed many changes addressed by Kritika. Primarily of the form:  - static final variables should be capita/lized  - Utility classes (all functions are static) should have 'Util'    or 'Helper' in the name.

+++ core/src/io/anuke/mindustry/editor/EditorTile.java
@@ -97,7 +97,7 @@ public class EditorTile extends Tile{
     }
 
     @Override
-    public void setOverlayID(short overlay){
+    public void setOverlayID(int overlay){
         if(state.is(State.playing)){
             super.setOverlayID(overlay);
             return;
